src/main/java/io/zenbydef/usertracker/service/oldpack/roleservice/userservice/UserServiceImpl.java
package io.zenbydef.usertracker.service.oldpack.roleservice.userservice;


import io.zenbydef.usertracker.io.entities.oldpack.User;
import io.zenbydef.usertracker.io.repositories.oldpack.UserRepository;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;


import java.util.List;
import java.util.Objects;

@Service
@Transactional
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;

    public UserServiceImpl(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public List<User> getUsers() {
        return userRepository.findAll();
    }

    @Override
    public void saveUser(User user) {
        userRepository.save(user);
    }

    @Override
    public User getUserById(Long id) {
        return userRepository.getOne(id);
    }

    @Override
    public void deleteUser(Long id) {
        userRepository.deleteById(id);
    }

    @Override
    public User findUserByName(String userName) {
        return userRepository.findUserByUsername(userName);
    }

    @Override
    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
        UserDetails foundUser = Objects.requireNonNull(userRepository.findUserByUsername(userName));
        foundUser.getAuthorities().size();
        return foundUser;
    }
}
